Edit article.md. Small section on truthy and falsy
The way that truthy and falsy works, especially how it evaluates against all falsy values, is a bit important so I thought it could be added to this section. Pretty much most of us turn to using truthy and falsy in conditional statements in place of traditional checks, and I felt that illustrating on it a bit more would be informative.