Update article.md
Hi Iliakan: 
I'm reading the text very carefully but still I find (as in my #335 submission) the same problem. 
What i'm trying to say is that you cannot test "integer property with this formula:
                 String(Math.trunc(Number("X")))
because:
alert(String(Math.trunc(Number("+49"))) === String(Math.trunc(Number("49")))) 
So there is a strict equality between Line 1 and Line 2, and so the above formula is not a good test for "integer property"!