Update solution.md
Replace 'this' in 'apply' call with empty object. Figured it was easier to bring up issue w/ PR than as a discussion.

This would be just for clarity, since, as far as I can tell, 'this' will always be 'undefined' in the returned function, but having it there anyway could make one wonder if if might actually have a useful value (if the new Function prototype was used with an object method). From what I've been able to discern, it would not -- it would always be 'undefined' (in strict mode, of course).

Same example I was wrong about before -- am I wrong again?

And, if I'm not, IS there a way to make the prototype wrapper work with an object method that uses 'this'?