confusing about working of  reduce method
On the table, there is a 'result' value for every calls of callback arrow function. But 'result' variable is not defined until end of 'reduce'  method execution. Because 'result' variable is in left hand side of assignment operator and assignment process will be done after 'reduce' method finished. There is a confusion because of naming convention.