conflict of function name and variable name
In the snippets, firstly g is a name of nested function and then g is a variable which carries address of the nested function. And tutorial talking about lexical environment of g and reach-ability of g, but which one ? Using "g" for different purpose is very bad idea. This situation is very good example of "ninja coding"!