to avoid override built-in properties
In the snippet we aren't using 'var' keyword so, it looks like modern JS codding standard. we block the 'var' danger but there is a way to override built-in properties in window object like Math object and this is very dangerous for project. To avoid this situation we need to check for "is property name in use?".