constructor function instead of constructor method
constructor methods make us recall "the constructor method in class declaration structure". But in here you show that the function has prototype object and in there is a attribute "constructor" which is a reference to the function itself. So we learn this pattern as a "constructor function" in this article:

https://javascript.info/function-prototype