not proved features of class properties
"The property name is not placed into User.prototype. Instead, it is created by new before calling the constructor, it’s a property of the object itself."

you are talking about some features of "class properties" but the example code doesn't show these features. I changed the example code to show that defining property out of the constructor method is different from defining method in class structure.