change about how some code "waits" for something
I can't think of a better way... somehow the idea of some code "waiting" for something is hard to think about. Is it by some kind of magically `.wait()` or by some kind of `sleep()` call?  I think the mechanism is: it just exists and waits to be called. I can't think of a better way to  put "wait for it" and "wait to be called" together.