TypedArray only represents other classes.
This may not be the best way to explain it, but I was mislead by the way that TypedArray was included in code examples as if it were an actual class. The list is a little unwieldy, but I don’t see anywhere else in this article that lists them all.