Fix link rot in 1-js/06-advanced-functions/02-rest-parameters-spread
There's a link that references a section about Object.assign in 1-js/04-object-basics/01-object
but the actual section about Object.assign is in 1-js/04-object-basics/02-object-copy.

I followed the advice in AUTHORING.md and didn't include the domain.  However, I also
used the # character to reference the specific section.  I'm not sure if that will
work or not.