Comma place
JUST a suggestion.  
I'm not sure about the correct use of the colon either.
The "correction" may be trivial (even wrong) for English. 
I HAD to do this change in the insanely long resulting sentence in Spanish. I asked Fundéu (Spanish academy) and things are more clear now.

Wrong about being wrong about being wrong.
I was studying
https://owl.purdue.edu/owl/general_writing/punctuation/commas/extended_rules_for_commas.html
Rules are different from Spanish,  not THAT different.  
So the suggestion. Again, it may be wrong.