declaring variable for Promise object can be danger
Declaring "let promise" variable can be dangerous. Because assignable variables makes asynchronous code difficult to debug. We can't know the promise object is resolved or still pending and  we can reassign the promise variable. So when did the promise resolved and which promise object is owner of result value in consuming code.