undefined instead of null for empty argument
Using undefined instead of null is more convenient for pass empty argument to a function. For the  example snippet is every thing is okay it works but in general we must send undefined valued argument to trigger "default parameter value" mechanism. For the example code we can pass any uncallable value instead of null, but undefined is more convenient for the purpose.