Discourage use of automatic variable creation
I've added explicit declarations of variables and not relied on automatic variable creation as it is [not the best of practices] (https://stackoverflow.com/a/6853267/362445) and not possible in IE5 strict mode. It also has different behaviour in IE vs other browsers. 

I've also added `<lable>` elements where it makes sense to do so, for semantic purposes, and as it is also considered best practice, and improves accessibility.