class syntax vs object literal
To tell the truth, the notations are not fully identical. There are subtle differences related to object inheritance (to be covered later), but for now they do not matter. In almost all cases the shorter syntax is preferred.

I thing this paragraph talks about prototypal inheritance but when we work with object literal there is no difference between  sayHi: function() { and sayHi() {  booth syntaxes add property as own property of object itself, does not add property to object's prototype like es6 class syntax.