Update prop descriptors article
Deletes info that it's possible to assign get/set for non-configurable accessor property if absent.

Maybe I just misunderstood. Could you please add some example about it if I'm wrong.

Please check this example.
https://codepen.io/artemtropanets/pen/bGqPKYx
If we have only a getter before setting 'configurable' to 'false' - then setter is 'undefined' and it's not possible to overwrite it later.
If we don't have get/set before changing 'configurable' then it's data property and it's not possible to change it to accessor property.