Fixing the unintentional behavior of the slider
When reading the article "Pointer events", I came across just a small bug.  When the pointer capture is set to the thumb, all consequent events including "pointermove" are retargeted back to the thumb. However, when the capture is lost, event listener for "pointermove" remains installed on the thumb element.  As a result, we can move inadvertently that thumb when we simply moving our mouse over it. To prevent its unintentional behavior, I suggest that we store capture pointer's id in a separate variable as long as capture persists.