Fix language errors and improve style
Technologies don't really "play" together, but they can "work" together.
It is best to avoid words such as "crippled", when "restricted" or "limited" are perfectly acceptable.
Change HTTP to HTTPS in the URLs in the paragraph above the image as HTTPS is in the image. I thought of changing anysite.com to javascript.info in this paragraph to match the image, but I decided not to. I may raise this later as a PR for you to decide.