Update commens in few code snippets
In the Non-writable and Non-configurable section, the comments in few of the code snippets says 'Error' but that is only the case when the 'strict' mode is enabled. I understand that it is also mentioned but I felt it should be reflected in the code snippet as well so that the reader is absolutely clear.