Fix Typos
I changed "parenthesis" to "parentheses" since "parenthesis" is singular and "parentheses" is plural and the sentence uses the plural version of the word. I changed "layed" to "laid" since "layed" is not the correct way to spell the past tense of "lay." Finally, I changed "setup" to "set up" since "setup" is a noun and "set up" is a verb and the verb should be used here.