tiny changes: spelling and consistency
Spelling:
`optimzed` -> `optimized`

Consistency: 
`Import *` -> `Import "*"` 

For consistency with `Import "as" and `Export "as"`.
Also -- as-is it looks like `Import *` is just `Import` with a footnote. It had me confused for several minutes.

NB: This shouldn't affect links, since neither `*` nor quotes are reflected in the link fragment. The URL is simply `/import-export#import` either way.