Update article.md - comment on skipping arguments
following the lessons, this is the first time we see skipping of optional variables. So I added an extra comment to point that out explicitly. 

also i guess we always need to refer to documents to know what a method's optional argument allows or considered rejected values - but it seems that passing undefined is the best approach to fallback internal default values (if implemented).